market_signals/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
apscheduler>=3.10.0
yfinance>=0.2.36
requests>=2.31.0
twilio>=9.0.0
python-dotenv>=1.0.0
pandas>=2.1.0
ta>=0.11.0
gunicorn>=22.0.0
pytz>=2024.1
